+++ packages/cinterop/src/commonMain/kotlin/io/realm/kotlin/internal/interop/sync/ProtocolErrorCode.kt
@@ -20,7 +20,7 @@ package io.realm.kotlin.internal.interop.sync
  * Wrapper for C-API `realm_sync_errno_client`.
  * See https://github.com/realm/realm-core/blob/master/src/realm.h#L3214
  */
-expect enum class ProtocolClientErrorCode {
+expect enum class ProtocolClientErrorCode : CodeDescription {
     RLM_SYNC_ERR_CLIENT_CONNECTION_CLOSED,
     RLM_SYNC_ERR_CLIENT_UNKNOWN_MESSAGE,
     RLM_SYNC_ERR_CLIENT_BAD_SYNTAX,
@@ -53,11 +53,8 @@ expect enum class ProtocolClientErrorCode {
     RLM_SYNC_ERR_CLIENT_HTTP_TUNNEL_FAILED,
     RLM_SYNC_ERR_CLIENT_AUTO_CLIENT_RESET_FAILURE;
 
-    // Public visible description of the enum value
-    public val description: String
-
     companion object {
-        fun fromInt(nativeValue: Int): ProtocolClientErrorCode
+        internal fun of(nativeValue: Int): ProtocolClientErrorCode?
     }
 }
 
@@ -65,7 +62,7 @@ expect enum class ProtocolClientErrorCode {
  * Wrapper for C-API `realm_sync_errno_connection`.
  * See https://github.com/realm/realm-core/blob/master/src/realm.h#L2942
  */
-expect enum class ProtocolConnectionErrorCode {
+expect enum class ProtocolConnectionErrorCode : CodeDescription {
     RLM_SYNC_ERR_CONNECTION_CONNECTION_CLOSED,
     RLM_SYNC_ERR_CONNECTION_OTHER_ERROR,
     RLM_SYNC_ERR_CONNECTION_UNKNOWN_MESSAGE,
@@ -82,11 +79,8 @@ expect enum class ProtocolConnectionErrorCode {
     RLM_SYNC_ERR_CONNECTION_SWITCH_TO_FLX_SYNC,
     RLM_SYNC_ERR_CONNECTION_SWITCH_TO_PBS;
 
-    // Public visible description of the enum value
-    public val description: String
-
     companion object {
-        fun fromInt(nativeValue: Int): ProtocolConnectionErrorCode
+        internal fun of(nativeValue: Int): ProtocolConnectionErrorCode?
     }
 }
 
@@ -94,7 +88,7 @@ expect enum class ProtocolConnectionErrorCode {
  * Wrapper for C-API `realm_sync_errno_session`.
  * See https://github.com/realm/realm-core/blob/master/src/realm.h#L2960
  */
-expect enum class ProtocolSessionErrorCode {
+expect enum class ProtocolSessionErrorCode : CodeDescription {
     RLM_SYNC_ERR_SESSION_SESSION_CLOSED,
     RLM_SYNC_ERR_SESSION_OTHER_SESSION_ERROR,
     RLM_SYNC_ERR_SESSION_TOKEN_EXPIRED,
@@ -126,10 +120,7 @@ expect enum class ProtocolSessionErrorCode {
     RLM_SYNC_ERR_SESSION_INITIAL_SYNC_NOT_COMPLETED,
     RLM_SYNC_ERR_SESSION_WRITE_NOT_ALLOWED;
 
-    // Public visible description of the enum value
-    public val description: String
-
     companion object {
-        fun fromInt(nativeValue: Int): ProtocolSessionErrorCode
+        internal fun of(nativeValue: Int): ProtocolSessionErrorCode?
     }
 }