+++ packages/cinterop/src/commonMain/kotlin/io/realm/kotlin/internal/interop/sync/ServiceErrorCode.kt
@@ -20,7 +20,7 @@ package io.realm.kotlin.internal.interop.sync
  * Wrapper for C-API `realm_app_errno_service`.
  * See https://github.com/realm/realm-core/blob/master/src/realm.h#L2559
  */
-expect enum class ServiceErrorCode {
+expect enum class ServiceErrorCode : CodeDescription {
     RLM_APP_ERR_SERVICE_MISSING_AUTH_REQ,
     RLM_APP_ERR_SERVICE_INVALID_SESSION,
     RLM_APP_ERR_SERVICE_USER_APP_DOMAIN_MISMATCH,
@@ -74,10 +74,7 @@ expect enum class ServiceErrorCode {
     RLM_APP_ERR_SERVICE_UNKNOWN,
     RLM_APP_ERR_SERVICE_NONE;
 
-    // Public visible description of the enum value
-    public val description: String
-
     companion object {
-        fun fromInt(nativeValue: Int): ServiceErrorCode
+        internal fun of(nativeValue: Int): ServiceErrorCode?
     }
 }