+++ packages/cinterop/src/darwin/kotlin/io/realm/kotlin/internal/interop/sync/AppErrorCategory.kt
@@ -18,26 +18,21 @@ package io.realm.kotlin.internal.interop.sync
 
 import realm_wrapper.realm_app_error_category
 
-actual enum class AppErrorCategory(actual val description: String, val nativeValue: realm_app_error_category) {
-    RLM_APP_ERROR_CATEGORY_HTTP("Http", realm_app_error_category.RLM_APP_ERROR_CATEGORY_HTTP),
-    RLM_APP_ERROR_CATEGORY_JSON("Json", realm_app_error_category.RLM_APP_ERROR_CATEGORY_JSON),
-    RLM_APP_ERROR_CATEGORY_CLIENT("Client", realm_app_error_category.RLM_APP_ERROR_CATEGORY_CLIENT),
-    RLM_APP_ERROR_CATEGORY_SERVICE("Service", realm_app_error_category.RLM_APP_ERROR_CATEGORY_SERVICE),
-    RLM_APP_ERROR_CATEGORY_CUSTOM("Custom", realm_app_error_category.RLM_APP_ERROR_CATEGORY_CUSTOM);
+actual enum class AppErrorCategory(
+    override val description: String,
+    override val nativeValue: Int
+) : CodeDescription {
+    RLM_APP_ERROR_CATEGORY_HTTP("Http", realm_app_error_category.RLM_APP_ERROR_CATEGORY_HTTP.value.toInt()),
+    RLM_APP_ERROR_CATEGORY_JSON("Json", realm_app_error_category.RLM_APP_ERROR_CATEGORY_JSON.value.toInt()),
+    RLM_APP_ERROR_CATEGORY_CLIENT("Client", realm_app_error_category.RLM_APP_ERROR_CATEGORY_CLIENT.value.toInt()),
+    RLM_APP_ERROR_CATEGORY_SERVICE("Service", realm_app_error_category.RLM_APP_ERROR_CATEGORY_SERVICE.value.toInt()),
+    RLM_APP_ERROR_CATEGORY_CUSTOM("Custom", realm_app_error_category.RLM_APP_ERROR_CATEGORY_CUSTOM.value.toInt());
 
     actual companion object {
 
-        actual fun fromInt(nativeValue: Int): AppErrorCategory {
-            for (value in values()) {
-                if (value.nativeValue.value.toInt() == nativeValue) {
-                    return value
-                }
+        internal actual fun of(nativeValue: Int): AppErrorCategory? =
+            values().firstOrNull { value ->
+                value.nativeValue == nativeValue
             }
-            error("Unknown app error category: $nativeValue")
-        }
-
-        internal fun of(nativeValue: realm_app_error_category): AppErrorCategory {
-            return fromInt(nativeValue.value.toInt())
-        }
     }
 }