+++ packages/cinterop/src/darwin/kotlin/io/realm/kotlin/internal/interop/sync/SyncErrorCodeCategory.kt
@@ -18,31 +18,21 @@ package io.realm.kotlin.internal.interop.sync
 
 import realm_wrapper.realm_sync_error_category
 
-actual enum class SyncErrorCodeCategory(actual val description: String, val nativeValue: realm_sync_error_category) {
-    RLM_SYNC_ERROR_CATEGORY_CLIENT("Client", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_CLIENT),
-    RLM_SYNC_ERROR_CATEGORY_CONNECTION("Connection", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_CONNECTION),
-    RLM_SYNC_ERROR_CATEGORY_SESSION("Session", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_SESSION),
-    RLM_SYNC_ERROR_CATEGORY_SYSTEM("System", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_SYSTEM),
-    RLM_SYNC_ERROR_CATEGORY_UNKNOWN("Unknown", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_UNKNOWN);
+actual enum class SyncErrorCodeCategory(
+    override val description: String,
+    override val nativeValue: Int
+) : CodeDescription {
+    RLM_SYNC_ERROR_CATEGORY_CLIENT("Client", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_CLIENT.value.toInt()),
+    RLM_SYNC_ERROR_CATEGORY_CONNECTION("Connection", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_CONNECTION.value.toInt()),
+    RLM_SYNC_ERROR_CATEGORY_SESSION("Session", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_SESSION.value.toInt()),
+    RLM_SYNC_ERROR_CATEGORY_SYSTEM("System", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_SYSTEM.value.toInt()),
+    RLM_SYNC_ERROR_CATEGORY_UNKNOWN("Unknown", realm_sync_error_category.RLM_SYNC_ERROR_CATEGORY_UNKNOWN.value.toInt());
 
     actual companion object {
 
-        actual fun fromInt(nativeValue: Int): SyncErrorCodeCategory {
-            for (value in values()) {
-                if (value.nativeValue.value.toInt() == nativeValue) {
-                    return value
-                }
+        internal actual fun of(nativeValue: Int): SyncErrorCodeCategory? =
+            values().firstOrNull { value ->
+                value.nativeValue == nativeValue
             }
-            error("Unknown sync error code category: $nativeValue")
-        }
-
-        fun of(nativeValue: realm_sync_error_category): SyncErrorCodeCategory {
-            for (value in values()) {
-                if (value.nativeValue == nativeValue) {
-                    return value
-                }
-            }
-            error("Unknown sync error code category: $nativeValue")
-        }
     }
 }